DroneLauncher & final for meet
Set drone launcher values, commented out code for telemetry that was causing errors and took out arming the position because it would take less fin tuning. Made the there have to be 2 inputs for the drone launch(to avoid accidental launches )
A close claw function for starting at the beginning of auto.